Moved MetadataApp to legacy
