+++ src/components/Application/Buttons/Create.tsx
@@ -6,8 +6,8 @@ import DatasetLinkedIcon from '@mui/icons-material/DatasetLinked'
 import CreateNewFolderIcon from '@mui/icons-material/CreateNewFolder'
 import HistoryEduIcon from '@mui/icons-material/HistoryEdu'
 import AddLinkIcon from '@mui/icons-material/AddLink'
-import DropdownButton from '../../../Parts/Buttons/Dropdown'
-import IconButton from '../../../Parts/Buttons/Icon'
+import DropdownButton from '../../Parts/Buttons/Dropdown'
+import IconButton from '../../Parts/Buttons/Icon'
 import DriveFolderUploadRounded from '@mui/icons-material/DriveFolderUploadRounded'
 import UploadFileRounded from '@mui/icons-material/UploadFileRounded'
 import { useStore } from '../store'
