+++ src/components/Application/Content.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react'
-import { ControllerProps } from '../../Controllers/Base'
-import File from '../../Controllers/File'
-import Metadata from '../../Controllers/Metadata'
-import Chart from '../../Controllers/Chart'
-import Table from '../../Controllers/Table'
-import Text from '../../Controllers/Text'
-import View from '../../Controllers/View'
-import Empty from '../../Parts/Empty'
-import Spinner from '../../Parts/Spinner'
+import { ControllerProps } from '../Controllers/Base'
+import File from '../Controllers/File'
+import Metadata from '../Controllers/Metadata'
+import Chart from '../Controllers/Chart'
+import Table from '../Controllers/Table'
+import Text from '../Controllers/Text'
+import View from '../Controllers/View'
+import Empty from '../Parts/Empty'
+import Spinner from '../Parts/Spinner'
 import { useStore } from './store'
 
 export default function Content() {
