+++ src/components/Application/Dialogs/CreateFolder.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import CreateNewFolderIcon from '@mui/icons-material/CreateNewFolder'
-import InputDialog from '../../../Parts/Dialogs/Input'
+import InputDialog from '../../Parts/Dialogs/Input'
 import { useStore, selectors } from '../store'
 
 export default function CreateFolderDialog() {
