+++ src/components/Application/Dialogs/DeleteFile.tsx
@@ -1,23 +1,23 @@
 import * as React from 'react'
 import DeleteIcon from '@mui/icons-material/Delete'
-import ConfirmDialog from '../../../Parts/Dialogs/Confirm'
+import ConfirmDialog from '../../Parts/Dialogs/Confirm'
 import { useStore } from '../store'
 
-export default function DeleteFolderDialog() {
+export default function DeleteFileDialog() {
   const path = useStore((state) => state.path)
-  const deleteFolder = useStore((state) => state.deleteFolder)
+  const deleteFile = useStore((state) => state.deleteFile)
   const updateState = useStore((state) => state.updateState)
   if (!path) return null
   return (
     <ConfirmDialog
       open={true}
-      title="Delete Folder"
+      title="Delete File"
       label="Delete"
       Icon={DeleteIcon}
       description={`You are deleting "${path}". Are you sure?`}
       onCancel={() => updateState({ dialog: undefined })}
       onConfirm={async () => {
-        await deleteFolder(path)
+        await deleteFile(path)
         updateState({ dialog: undefined })
       }}
     />
