+++ src/components/Application/Dialogs/DeleteFolder.tsx
@@ -1,23 +1,23 @@
 import * as React from 'react'
 import DeleteIcon from '@mui/icons-material/Delete'
-import ConfirmDialog from '../../../Parts/Dialogs/Confirm'
+import ConfirmDialog from '../../Parts/Dialogs/Confirm'
 import { useStore } from '../store'
 
-export default function DeleteFileDialog() {
+export default function DeleteFolderDialog() {
   const path = useStore((state) => state.path)
-  const deleteFile = useStore((state) => state.deleteFile)
+  const deleteFolder = useStore((state) => state.deleteFolder)
   const updateState = useStore((state) => state.updateState)
   if (!path) return null
   return (
     <ConfirmDialog
       open={true}
-      title="Delete File"
+      title="Delete Folder"
       label="Delete"
       Icon={DeleteIcon}
       description={`You are deleting "${path}". Are you sure?`}
       onCancel={() => updateState({ dialog: undefined })}
       onConfirm={async () => {
-        await deleteFile(path)
+        await deleteFolder(path)
         updateState({ dialog: undefined })
       }}
     />
