+++ src/components/Application/Dialogs/MoveFile.tsx
@@ -1,25 +1,25 @@
 import * as React from 'react'
-import ContentCopyIcon from '@mui/icons-material/ContentCopy'
-import InputDialog from '../../../Parts/Dialogs/Input'
+import CopyAllIcon from '@mui/icons-material/CopyAll'
+import InputDialog from '../../Parts/Dialogs/Input'
 import { useStore } from '../store'
 
-export default function CopyFileDialog() {
+export default function MoveFileDialog() {
   const path = useStore((state) => state.path)
-  const copyFile = useStore((state) => state.copyFile)
+  const moveFile = useStore((state) => state.moveFile)
   const updateState = useStore((state) => state.updateState)
   if (!path) return null
   return (
     <InputDialog
       open={true}
       value={path}
-      title="Copy File"
-      label="Copy"
-      Icon={ContentCopyIcon}
+      title="Move File"
+      label="Move"
+      Icon={CopyAllIcon}
       placholder="Enter a path"
-      description={`You are copying "${path}". Enter destination:`}
+      description={`You are moving "${path}". Enter destination:`}
       onCancel={() => updateState({ dialog: undefined })}
       onConfirm={async (toPath) => {
-        await copyFile(path, toPath)
+        await moveFile(path, toPath)
         updateState({ dialog: undefined })
       }}
     />
