+++ src/components/Application/Dialogs/MoveFolder.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import CopyAllIcon from '@mui/icons-material/CopyAll'
-import InputDialog from '../../../Parts/Dialogs/Input'
+import InputDialog from '../../Parts/Dialogs/Input'
 import { useStore } from '../store'
 
 export default function MoveFolderDialog() {
