+++ src/components/Application/Menu.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import InputAdornment from '@mui/material/InputAdornment'
-import SelectField from '../../Parts/Fields/Select'
-import * as menu from '../../Parts/Bars/Menu'
+import SelectField from '../Parts/Fields/Select'
+import * as menu from '../Parts/Bars/Menu'
 
 export default function Menu() {
   return (
