+++ src/components/Application/Status.tsx
@@ -2,12 +2,12 @@ import * as React from 'react'
 import Box from '@mui/material/Box'
 import { alpha, styled } from '@mui/material/styles'
 // import InputIcon from '@mui/icons-material/Input'
-import LightTooltip from '../../Parts/Tooltips/Light'
-import ReportChip from '../../Parts/Chips/Report'
-import ChartChip from '../../Parts/Chips/Chart'
-import ViewChip from '../../Parts/Chips/View'
-import ScriptChip from '../../Parts/Chips/Script'
-import TypeChip from '../../Parts/Chips/Type'
+import LightTooltip from '../Parts/Tooltips/Light'
+import ReportChip from '../Parts/Chips/Report'
+import ChartChip from '../Parts/Chips/Chart'
+import ViewChip from '../Parts/Chips/View'
+import ScriptChip from '../Parts/Chips/Script'
+import TypeChip from '../Parts/Chips/Type'
 import { useStore } from './store'
 
 export default function Status() {
