+++ src/components/Application/types.ts
