+++ src/metadata.ts
@@ -3,7 +3,6 @@ import '@fontsource/roboto/300.css'
 import '@fontsource/roboto/400.css'
 import '@fontsource/roboto/500.css'
 import '@fontsource/roboto/700.css'
-import Metadata from './components/Applications/Metadata'
 import Package from './components/Editors/Package'
 import Resource from './components/Editors/Resource'
 import Dialect from './components/Editors/Dialect'
@@ -11,4 +10,4 @@ import Schema from './components/Editors/Schema'
 import Report from './components/Parts/Report'
 import { render } from './render'
 
-export { Metadata, Package, Resource, Dialect, Schema, Report, render }
+export { Package, Resource, Dialect, Schema, Report, render }
